requirements.txt
astroid==2.5.8
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.1
colorama==0.4.4
dnspython==2.1.0
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
greenlet==1.1.0
gunicorn==20.0.4
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.20.3
pandas==1.2.4
pip==21.1.2
psycopg2-binary==2.9.1
pylint==2.8.3
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
setuptools==52.0.0.post20210125
six==1.16.0
soupsieve==2.2.1
splinter==0.14.0
SQLAlchemy==1.3.15
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.36.2
wincertstore==0.2
wrapt==1.12.1
